refactor: rename chunk ir -> chunk ops
