+++ src/etha/comm/__init__.py
@@ -2,11 +2,11 @@
 
 from .get_m2m_map import get_m2m_map
 from .comm_methods import m2m_communicate, gather_broadcast_communicate
-from .get_chunk_ir import map_to_chunk_ir
+from .get_chunk_ops import map_to_chunk_ops
 
 __all__ = [
     "get_m2m_map",
-    "map_to_chunk_ir",
+    "map_to_chunk_ops",
     "m2m_communicate",
     "gather_broadcast_communicate",
 ]
